+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Minor: Include link to killed player profile in PK notifier. (#246)
 - Minor: Add setting to exclude group name from GIM shared storage notifications. (#247)
 - Minor: Include relevant wiki links in rich embed content. (#243, #248)
 - Bugfix: Ignore notifications from new beta worlds. (#253)